Adjust version update suffix in java-sdk to avoid brew revision suffix match.
With pure numeric suffix (e.g., 1.7.0_67), PkgVersion incorrectly matches
version as '1.7.0' and revision as 67. This result in incorrect version
comparision in method '<=>' and the package is always marked 'outdated'.

We should not be using 'revision' label either because it is meant to
mark newer revision of the brew formula/package and not the version of
the actual binaries.